move decrypt/memory-mode to parent
