+++ src/crypto.rs
@@ -1,3 +1,2 @@
-pub mod decrypt;
 pub mod key;
 pub mod primitives;
